+++ packages/terra-arrange/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * Changed
+  * Upgrade tests to terra-functional-testing and WebDriverIO v6
+  * Removed unnecessary `wdio.conf.js` file
   * Updated Jest command
 
 ## 3.43.0 - (February 2, 2021)