+++ packages/terra-arrange/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+## 3.45.0 - (April 13, 2021)
+
+* Changed
+  * Minor dependency version bump
+
+## 3.44.0 - (April 6, 2021)
+
 * Changed
   * Upgrade tests to terra-functional-testing and WebDriverIO v6
   * Removed unnecessary `wdio.conf.js` file